Rework PeerListener to allow for crashes without memory leaks
